Using mock ProductCarouselItem for product-carousel tests and moving UI tests to product-carousel-item Providing type for Product Removing unnecessary config in product-carousel module as it goes unseen by the customer anyway SCSS for showing multiple CMS components in product-carousel-item
