+++ projects/storefrontstyles/scss/components/product/carousel/_product-carousel-item.scss
@@ -2,5 +2,9 @@
   .actions {
     width: 60%;
     margin: auto;
+
+    > *:not(:first-child) {
+      margin-bottom: 4px;
+    }
   }
 }
